Move .proto files up for reuse
