+++ build.json
@@ -578,8 +578,8 @@
       "build": "private",
       "language": "c++",
       "src": [
-        "test/cpp/interop/empty.proto",
-        "test/cpp/interop/messages.proto",
+        "test/proto/empty.proto",
+        "test/proto/messages.proto",
         "test/proto/test.proto",
         "test/cpp/interop/client.cc",
         "test/cpp/interop/interop_client.cc"
@@ -613,8 +613,8 @@
       "build": "private",
       "language": "c++",
       "src": [
-        "test/cpp/interop/empty.proto",
-        "test/cpp/interop/messages.proto",
+        "test/proto/empty.proto",
+        "test/proto/messages.proto",
         "test/proto/test.proto",
         "test/cpp/interop/server.cc"
       ],
