+++ test/cpp/interop/server.cc
@@ -49,8 +49,8 @@
 #include <grpc++/status.h>
 #include <grpc++/stream.h>
 #include "test/proto/test.grpc.pb.h"
-#include "test/cpp/interop/empty.grpc.pb.h"
-#include "test/cpp/interop/messages.grpc.pb.h"
+#include "test/proto/empty.grpc.pb.h"
+#include "test/proto/messages.grpc.pb.h"
 #include "test/cpp/interop/server_helper.h"
 #include "test/cpp/util/test_config.h"
