+++ test/proto/empty.proto
