+++ test/proto/test.proto
@@ -30,10 +30,11 @@
 
 // An integration test service that covers all the method signature permutations
 // of unary/streaming requests/responses.
+
 syntax = "proto2";
 
-import "test/cpp/interop/empty.proto";
-import "test/cpp/interop/messages.proto";
+import "test/proto/empty.proto";
+import "test/proto/messages.proto";
 
 package grpc.testing;
